check was redisigned, fixed api responses handling

+++ app/src/main/java/com/betkey/network/models/CouponSettings.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 class CouponSettings (
     @SerializedName("pool")
-    var pool: Int? = null,
+    var pool: Double? = null,
 
     @SerializedName("profit")
     var profit: Int? = null,
